Protecting against calls to wiping entire directories
First, the clean=True is removed from the ForcedDirectoryChanger.
The purpose being that option is really what causes the issue,
while this may break some people's workflows, having code around
that recursively deletes directories is pretty scary, and those
workflows should change.

Second, the pre-existing cleanPath method was moved to a better
location and is now used in place of all bare calls to
shutil.rmtree. It is scary that such a method was called in 8
different places in the repo, using a bare call with no safety
checks. At the very least, such things should be wrapped in a
lot of safety catches.

Lastly, there is a whitelisted set of directories in cleanPath
that can be wiped. In an ideal world, the codebase and our
workflows would change to allow for only one: /.armi/ .
