+++ armi/utils/pathTools.py
@@ -21,6 +21,7 @@ import os
 import shutil
 import importlib
 import pathlib
+from time import sleep
 
 from armi import runLog
 
@@ -215,3 +216,72 @@ def moduleAndAttributeExist(pathAttr):
     except:
         return False
     return moduleAttributeName in userSpecifiedModule.__dict__
+
+
+def cleanPath(path):
+    """Recursively delete a path.
+
+    !!! careful with this !!! It can delete the entire cluster.
+
+    We add copious os.path.exists checks in case an MPI set of things is trying to delete everything at the same time.
+    Always check filenames for some special flag when calling this, especially
+    with full permissions on the cluster. You could accidentally delete everyone's work
+    with one misplaced line! This doesn't ask questions.
+
+    Safety nets include a whitelist of paths.
+
+    This makes use of shutil.rmtree and os.remove
+
+    Returns
+    -------
+    success : bool
+        True if file was deleted. False if it was not.
+
+    """
+    valid = False
+    if os.path.exists(path):
+        runLog.extra("Clearing all files in {}".format(path))
+    else:
+        runLog.extra("Nothing to clean in {}. Doing nothing. ".format(path))
+        return True
+
+    for validPath in [
+        ".armi",
+        "armiruns",
+        "failedruns",
+        "mc2run",
+        "mongoose",
+        "shufflebranches",
+        "snapshot",
+        "tests",
+    ]:
+        if validPath in path.lower():
+            valid = True
+
+    if not valid:
+        raise Exception(
+            "You tried to delete {0}, but it does not seem safe to do so.".format(path)
+        )
+
+    for i in range(3):
+        try:
+            if os.path.exists(path) and os.path.isdir(path):
+                shutil.rmtree(path)
+            elif not os.path.isdir(path):
+                # it's just a file. Delete it.
+                os.remove(path)
+        except:
+            if i == 2:
+                pass
+            # in case the OS is behind or something
+            sleep(0.1)
+
+        sleep(0.3)
+        if not os.path.exists(path):
+            break
+        sleep(0.3)
+
+    if os.path.exists(path):
+        return False
+    else:
+        return True
